Heavily refactored the build script. Implemented header files for the kernel. Added A kernel panic system (1 being lowest danger)
