+++ src/circuitpanic.h
@@ -1,0 +1,7 @@
+#ifndef CURCUITPANIC_H
+#define CURCUITPANIC_H
+#include "kernel.h"
+
+void CPanic1(const char* message);
+
+#endif // CURCUITPANIC_H
