+++ src/kernel.c
@@ -2,6 +2,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include "kernel.h"
+#include "circuitpanic.h"
+
 /* Check if the compiler thinks you are targeting the wrong operating system. */
 #if defined(__linux__)
 #error "Targeting wrong system!"
@@ -12,25 +15,6 @@
 #error "An ix86-elf compiler needs to be used!"
 #endif
 
-/* Hardware text mode color constants. */
-enum vga_color {
-	VGA_COLOR_BLACK = 0,
-	VGA_COLOR_BLUE = 1,
-	VGA_COLOR_GREEN = 2,
-	VGA_COLOR_CYAN = 3,
-	VGA_COLOR_RED = 4,
-	VGA_COLOR_MAGENTA = 5,
-	VGA_COLOR_BROWN = 6,
-	VGA_COLOR_LIGHT_GREY = 7,
-	VGA_COLOR_DARK_GREY = 8,
-	VGA_COLOR_LIGHT_BLUE = 9,
-	VGA_COLOR_LIGHT_GREEN = 10,
-	VGA_COLOR_LIGHT_CYAN = 11,
-	VGA_COLOR_LIGHT_RED = 12,
-	VGA_COLOR_LIGHT_MAGENTA = 13,
-	VGA_COLOR_LIGHT_BROWN = 14,
-	VGA_COLOR_WHITE = 15,
-};
 
 static inline uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg) 
 {
@@ -117,4 +101,11 @@ void kernel_main(void)
 	// Initialize terminal & write string
 	terminal_initialize();
 	terminal_writestring("Wireframe kernel v0.1.0 ALPHA\n");
+
+	// Testing a kernel panic function. Was up till 1:17 working on this, so I'm going to bed.
+	// it would be so sigma if you accepted this pull request UwU
+	
+	
+
+	// CPanic1("Kernel panic: Unhandled exception\n");
 }
